requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
streamlit>=1.35.0
openai>=1.12.0
faiss-cpu>=1.7.4
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
typer>=0.9.0
rich>=13.0.0
loguru>=0.7.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
black>=23.0.0
mypy>=1.8.0
flake8>=6.0.0
bandit>=1.7.5
pre-commit>=3.6.0

# Optional dependencies
watchdog>=3.0.0  # For file watching
psutil>=5.9.0    # For system monitoring
requests>=2.31.0 # For HTTP requests

# Windows-specific: Use pre-compiled packages
numpy>=1.24.0  # Ensure compatible numpy version
scipy>=1.10.0  # Pre-compiled scipy